pyproject.toml
[tool.poetry]
name = "similarity_of_texts"
version = "0.1.0"
description = ""
authors = ["cyber jq"]

[tool.poetry.dependencies]
python = "^3.8"
gensim = "^4.2.0"
python-docx = "^0.8.11"
pywin32 = "^304"
nltk = "^3.7"
pymorphy2 = "^0.9.1"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
